Add travis builds
