+++ .travis.yml
@@ -1,0 +1,6 @@
+language: csharp
+solution: ./src/BOG.SwissArmyKnife.sln
+mono: none
+dotnet: 3.1
+script:
+- dotnet build ./src/BOG.SwissArmyKnife.sln
